rename state interface files
