+++ src/app/products/state/product.reducer.ts
@@ -1,4 +1,4 @@
-import { ProductState } from './product-state';
+import { ProductState } from './product.state';
 
 const initialState: ProductState = {
 	showProductCode: true,
